Move error types to dedicated mod
